refactor : AttendanceManager의 reward tag관련 메소드를 분리함
- AttendanceManager의 reward tag 관련 메소드를 AttendanceRewardManager로 분리했습니다.
- 관련 테스트 케이스를 추가했습니다.
