+++ waldreg/service-layer/attendance/src/main/java/org/waldreg/attendance/management/DefaultAttendanceManager.java
@@ -15,9 +15,7 @@ import org.waldreg.attendance.management.dto.AttendanceStatusChangeDto;
 import org.waldreg.attendance.management.dto.AttendanceTargetDto;
 import org.waldreg.attendance.management.dto.AttendanceUserDto;
 import org.waldreg.attendance.management.spi.AttendanceRepository;
-import org.waldreg.attendance.management.spi.AttendanceRewardRepository;
 import org.waldreg.attendance.management.spi.UserExistChecker;
-import org.waldreg.attendance.valid.AttendanceIdentifyValidable;
 import org.waldreg.attendance.rule.valid.AttendanceDateValidator;
 import org.waldreg.attendance.type.AttendanceType;
 import org.waldreg.attendance.valid.AttendanceTargetValidable;
@@ -29,20 +27,14 @@ public class DefaultAttendanceManager implements AttendanceManager, AttendanceTa
     private final UserExistChecker userExistChecker;
     private final AttendanceRepository attendanceRepository;
     private final AttendanceDateValidator attendanceDateValidator;
-    private final AttendanceIdentifyValidable attendanceIdentifyValidable;
-    private final AttendanceRewardRepository attendanceRewardRepository;
 
     @Autowired
     public DefaultAttendanceManager(UserExistChecker userExistChecker,
                                     AttendanceRepository attendanceRepository,
-                                    AttendanceDateValidator attendanceDateValidator,
-                                    AttendanceIdentifyValidable attendanceIdentifyValidable,
-                                    AttendanceRewardRepository attendanceRewardRepository){
+                                    AttendanceDateValidator attendanceDateValidator){
         this.userExistChecker = userExistChecker;
         this.attendanceRepository = attendanceRepository;
         this.attendanceDateValidator = attendanceDateValidator;
-        this.attendanceIdentifyValidable = attendanceIdentifyValidable;
-        this.attendanceRewardRepository = attendanceRewardRepository;
     }
 
     @Override
@@ -70,19 +62,6 @@ public class DefaultAttendanceManager implements AttendanceManager, AttendanceTa
     private void throwIfCannotFindUser(int id){
         if(!userExistChecker.isExistUser(id)){
             throw new UnknownUsersIdException(id);
-        }
-    }
-
-    @Override
-    public void setRewardTag(int rewardTagId, AttendanceType attendanceType){
-        throwIfRewardTagDoesNotExist(rewardTagId);
-        throwIfUnknownAttendanceType(attendanceType);
-        attendanceRewardRepository.setRewardTag(rewardTagId, attendanceType);
-    }
-
-    private void throwIfRewardTagDoesNotExist(int rewardTagId){
-        if(!attendanceRewardRepository.isRewardTagExist(rewardTagId)){
-            throw new UnknownRewardTagIdException(rewardTagId);
         }
     }
