+++ waldreg/service-layer/attendance/src/main/java/org/waldreg/attendance/reward/AttendanceRewardManager.java
@@ -1,0 +1,9 @@
+package org.waldreg.attendance.reward;
+
+import org.waldreg.attendance.type.AttendanceType;
+
+public interface AttendanceRewardManager{
+
+    void setRewardTag(int rewardTagId, AttendanceType attendanceType);
+
+}
